Market Manager and World Update
A market manager that will place markets at the start of a game has been
implemented.

The world has been changed from the graphic sprites to a line detailed
sprite.

A spawn notifier that shows the player a new mission has appeared has
been implemented.

A method was added to Continent that will return a random point on that
continent.